third_party/blink/web_tests/platform/generic/fast/shapes/shape-outside-floats/shape-outside-rounded-boxes-002-expected.txt

PASS elementRect('a').top is 0
PASS elementRect('a').right is 200
PASS elementRect('b').top is 20
PASS Math.floor(elementRect('b').right) is 105
PASS elementRect('c').top is 40
PASS elementRect('c').right is 100
PASS elementRect('d').top is 60
PASS elementRect('d').right is 100
PASS elementRect('e').top is 80
PASS Math.floor(elementRect('e').right) is 105
PASS elementRect('f').top is 100
PASS elementRect('f').right is 200
PASS successfullyParsed is true

TEST COMPLETE
